CRM/functions/exotel_call_webhook/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.11.0
google-cloud-secret-manager>=2.16.0
vertexai>=1.0.0
requests>=2.31.0
